requirements.txt
absl-py                  0.15.0
astunparse               1.6.3
cachetools               4.2.4
certifi                  2022.12.7
charset-normalizer       3.4.0
click                    8.1.8
cloudpickle              2.2.1
colorama                 0.4.6
coloredlogs              15.0.1
cycler                   0.11.0
dgl                      2.0.0
dgl-cu111                0.6.1
et-xmlfile               1.1.0
Flask                    2.2.5
flatbuffers              1.12
future                   1.0.0
gast                     0.3.3
google-auth              1.35.0
google-auth-oauthlib     0.4.6
google-pasta             0.2.0
grpcio                   1.32.0
gym                      0.26.2
gym-notices              0.0.8
h5py                     2.10.0
humanfriendly            10.0
idna                     3.10
importlib-metadata       6.7.0
iso8601                  2.1.0
itsdangerous             2.1.2
Jinja2                   3.1.6
Keras-Preprocessing      1.1.2
kiwisolver               1.4.5
Markdown                 3.4.4
markdown-it-py           2.2.0
MarkupSafe               2.1.5
matplotlib               3.2.1
mdurl                    0.1.2
mpmath                   1.3.0
networkx                 2.6.3
numpy                    1.21.6
oauthlib                 3.2.2
onnx                     1.14.1
onnxruntime              1.14.1
onnxsim                  0.4.36
opencv-python            4.2.0.34
openpyxl                 3.1.3
opt-einsum               3.3.0
packaging                24.0
pandas                   1.3.5
Pillow                   8.2.0
pip                      22.3.1
portalocker              2.7.0
protobuf                 3.20.3
psutil                   7.0.0
pyasn1                   0.5.1
pyasn1-modules           0.3.0
pygame                   2.6.1
Pygments                 2.17.2
pyparsing                3.1.4
pyreadline               2.1
python-dateutil          2.9.0.post0
pytz                     2025.1
pywin32                  308
PyYAML                   6.0.1
requests                 2.31.0
requests-oauthlib        2.0.0
rich                     13.8.1
rsa                      4.9
scipy                    1.4.1
serial                   0.0.97
setuptools               65.6.3
six                      1.15.0
sympy                    1.10.1
tensorboard              2.11.2
tensorboard-data-server  0.6.1
tensorboard-plugin-wit   1.8.1
tensorflow-estimator     2.4.0
tensorflow-gpu           2.4.0
tensorflow-gpu-estimator 2.2.0
termcolor                1.1.0
tf2onnx                  1.16.1
torch                    1.13.1
torchdata                0.5.1
tqdm                     4.46.1
typing_extensions        4.7.1
urllib3                  2.0.7
Werkzeug                 2.2.3
wheel                    0.38.4
wincertstore             0.2
wrapt                    1.12.1
zipp                     3.15.0